User Registration

+++ studybud/base/views.py
@@ -1,7 +1,10 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.db.models import Q
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth.models import User
 
 from .models import Room, Topic
@@ -9,8 +12,13 @@ from .forms import RoomForm
 
 
 def login_page(request):
+    page = 'login'
+    
+    if request.user.is_authenticated:
+        return redirect('home')
+    
     if request.method == 'POST':
-        username = request.POST.get("username")
+        username = request.POST.get("username").lower()
         password = request.POST.get("password")
     
         try:
@@ -27,12 +35,28 @@ def login_page(request):
         else:
             messages.error(request, 'Username or password does not exist. ')
         
-    context = {}
+    context = {'page':page}
     return render(request, 'base/login_register.html', context)
 
 def logout_user(request):
     logout(request)
     return redirect('home')
+
+def register_page(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username  = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "An error occured during registration.")
+
+    return render(request, 'base/login_register.html', {'form':form})
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -51,6 +75,7 @@ def room(request, pk):
     room = Room.objects.get(id=pk)
     return render(request, 'base/room.html', {'room': room})
 
+@login_required(login_url="login")
 def create_room(request):
     form = RoomForm()
     
@@ -63,9 +88,13 @@ def create_room(request):
     context = {"form": form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url="login")
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host :
+        return HttpResponse("You are not allowed here!")
     
     if request.method == 'POST':
         form = RoomForm(request.POST,instance=room)
@@ -76,8 +105,13 @@ def update_room(request, pk):
     context = {"form": form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url="login")
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!")
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
